[ENH] ForecastingOptunaSearchCV for hyper-parameter tuning for forecasting (#6630)
#### Reference Issues/PRs
Closes #6618. See also #4188 


#### What does this implement/fix? Explain your changes.

While sktime provides a ForecastingGridSearchCV,
ForecastingRandomizedSearchCV and ForecastingSkoptSearchCV for grid
search, there's a need for more sophisticated and efficient
hyper-parameter optimization methods. The lack of integration with
advanced optimization libraries like Optuna limits the ability to
perform more complex and efficient hyper-parameter searches.
We propose the addition of a ForecastingOptunaSearchCV class to sktime's
forecasting module. This class would leverage Optuna for hyper-parameter
optimization, providing a more efficient and flexible way to find the
best parameters for forecasting models. The interface is similar to the
existing ForecastingGridSearchCV, allowing users to easily switch
between grid search and Optuna-based optimization without significant
changes to their code.


#### Does your contribution introduce a new dependency? If yes, which
one?


Adds a dependency on Optuna. 

#### Did you add any tests for the change?
We added test_optuna in
/sktime/sktime/forecasting/model_selection/tests/test_tune.py