[ENH] refactor and bugfixes for environment checker utilities (#6719)
This PR refactors and fixes some silent bugs in the environment checker
utiliits.

Refactor:

* checking of presence of individual packages and their versions is now
factored out and `lru_cache`-d into a utility `_get_pkg_version`. This
avoids checking presence or version of the same package twice.
* multiplied boilerplate for raising exception, warning, or nothing
based on `severity` args is factored out into a single private method,
`_raise_at_severity`

Bugfix:
* the refactor of `_raise_at_severity` has revealed that for multiple
functions, the return in the case of `severity="warning"` was incorrect
(`True` even if the package was missing or incompatible). This is not a
serious bug as that case was no longer used in the code base. Either
way, it is now fixed.