[BUG] fix polymorphic estimators missing in estimator overview, e.g., `pytorch-forecasting` forecasters (#6803)
Fixes #6801, the missing `pytorch-forecasting` forecasters in the
forecaster estimator overview. This was a consequence of a more general
bug where subsetting specifically for the estimator overview was
incorrect for polymorphic, i.e., multi-type estimators.

The reason for the bug was that the estimator overview subsetting
reimplemented type subsetting logic in `all_estimators`, and subsetted
only to the first type in case there were multiple.
(and `pytorch-forecasting` forecasters are polymorphic, since currently
global forecasting is implemented as a separate type, from "ordinary"
forecasting)

We should refactor and deduplicate the code in `docs/conf.py`, but for
now I've fixed the logic mistake.