[BUG] fix `MSTL` inverse transform and use in forecasting pipeline (#6825)
Fixes https://github.com/sktime/sktime/issues/6397, and more generally
ensures that `MSTL` can be used in forecasting pipelines as expected.
Also improves documentation, and fixes some unreported bugs in `MSTL`.

* `return_components` would not return components that sum to the
original as expected. This has been fixed. I also consider the fact that
the return in the case `return_components=True` included "transformed",
which is "trend" plus "residual", a bug, since unexpected by the user.
* The docstring was incorrect in describing the components that were
actually returned, in the case where `periods` were passed - not three
components with seasonal components summed up, but seasonal components
individually.
* `inverse_transform` did not work. This was also masked by the
boilerplate issue solved in #6824. This is now also tested, which also
covers the fix in #6824.
* The type of the `seasonal_` attribute was inconsistent across the
cases of passing `periods` or not, and this was also incorrectly
described in the docstring. The type has been rendered consistent
(inconsistency was due to an unreported bug), `pd.DataFrame`, and this
is now also described in the docstring.

Depends on https://github.com/sktime/sktime/pull/6824