[ENH] making `self._is_vectorized` access more defensive in `BaseTransformer` (#6863)
This PR makes the new access points for `self._is_vectorized` in
`BaseTransformer` more defensive, by using `getattr` with a default
instead of direct attribute access. The changes were introduced in
https://github.com/sktime/sktime/pull/6825 to fix a bug with
vectorization ov `inverse_transform`.

As @Abhay-Lejith has pointed out, the `SFA` transformer is failing as it
overrides `fit` but not `transform` (see
https://github.com/sktime/sktime/actions/runs/10147989805/job/28060326703?pr=6732).

This is not something an estimator should do, as it breaks the API
extension pattern, but in case downstream users copied this pattern, at
least they will not be impacted.

Also makes some minor changes to `SFA` transformer (setting the
attribute in `fit`), and `SFAfast` (docstring formatting).