[DOC] clarify column handling in docstring of `FourierFeatures` (#6834)
This PR clarifies the docstring of `FourierFeatures` about handling of
the original columns. Previously, this was not too clear from the
preamble, and only became evident after looking at all parameters.

This addition is to prevent the confusion in
https://github.com/sktime/sktime/issues/6674 - when the transformer is
used in isolation, a user might expect the original columns to be
retained. However, by default, the columns are dropped, to ensure the
expected behaviour when used as part of a pipeline.