[DOC] change links in documentation to refer to same version (#6841)
In most places in `docs/source`, links are hardcoded to either `latest`
(corresponding to `main` branch) or to `stable` (most recent tag). But
quite often users will probably prefer to see the documentation for the
same version they are browsing, assuming because they are using that
version at the momemt. This PR makes these changes, following
https://docs.readthedocs.io/en/stable/guides/cross-referencing-with-sphinx.html#the-doc-role.