[BUG] Refactor ADICVTransformer and fix CV calculation (#6757)
Fixes coefficient of variation calculation and refactors some of the
corresponding classification code. Also adds parameter for the method
used to calculate ADI. Updates tests to remove randomly-generated test
data and create reproducible tests instead.

#### Reference Issues/PRs
Fixes #6755 


#### What does this implement/fix? Explain your changes.
Addresses an incorrect denominator in the coefficient of variance
squared calculation. It also adds a parameter to allow the end user to
control the end behavior of the average demand interval calculation.