[BUG] changelog utility: fix termination condition to retrieve merged PR (#6920)
This PR fixes a bug in the changelog utility, namely the termination
condition to retrieve all merged PR since the last release.

The bug manifests as less PR being retrieved than merged.

The reason for this is the termination condition: the script retrieves
PR, sorted by "updated at", until a PR is encountered that was merged
before the release.

This terminates too early, in all cases where a merged PR before the
last release gets updated so it ends up more recent than the last page
that contains merged PR after the last release. The subsequent pages of
PRs are incorrectly cut off then.

The fix is simple, the termination condition should be when a PR is
encountered whose latest *update* was before the last release. Since PR
are ordered by when the latest update was, this ensures that all PRs
that got updated later than the release get checked. The set of PR that
got merged is a subset of these, so the condition is sufficient to check
all relevant PRs.