[ENH] improve test_global_forecasting_tag (#6929)
This PR fixes `test_global_forecasting_tag` to incorporate classes that
are not inherited by `_BaseGlobalForecaster` rather inherit a class that
inherits this. Current behaviour does not detect that the estimator
supports global forecasting if it does not directly inherit
`_BaseGlobalForecaster` and the global tests are not executed.

Problem was identified in https://github.com/sktime/sktime/pull/6928
where `LTSFLinearForecaster` inherits `BaseDeepNetworkPyTorch` which
inherits `_BaseGlobalForecaster` and global capability could not be
detected during tests.