[ENH] New estimator - RegularizedVAR (#6725)
This PR introduces a new `VARReduce` class, which implements a VAR-like
forecaster that reduces multivariate time series to tabular regression
and trains a regressor.
Any scikit-learn compatible regressor can be used; by default,
`LinearRegression` is used, making it behave like a traditional VAR.
Users can specify other regressors such as `ElasticNet` to introduce
regularization and potentially enhance performance with large datasets
or in the presence of multicollinearity.

#### Did you add any tests for the change?
Yes, I've tested the fitted parameters and forecasting results of this
new class versus a 'normal' VAR from `statsmodels` on a synthetic
multivariate data. The results, which are shown in an example notebook,
are almost identical.