[ENH] improvements to `BaseSeriesAnnotator` base class for anomaly, changepoint, segments (#7073)
This PR improves the structure of the `BaseSeriesAnnotator` class, with
the goal to:

* remove some obstacles that prevented the class from being easily
composable
* make the extender interface clearer

Changes: 

* the `task` and `learning_type` params in the init were set using the
class tag, which prevented dynamic setting in class, this is changed to
`set_tag`
* in the boilerplate, `self.task` was used instead of the tag - this is
changed in a move to use the tag system entirely. `self.task` is left
for downwards compatibility.
* defaults are implemented for `_predict_segments` and
`_predict_points`, this allows extenders to focus on implementing
`_predict` only.