[BUG] fix probabilistic forecasts if only `_predict_var` is implemented (#7153)
Forecasters probabilistic forecasts would fail if only `_predict_var`
was implemented, as the dispatching logic was not correctly worked out.
The bug was undetected and unreported, as #6424 seems to add the first
such example.

I thought it was, but it was correct only in `sdkpro`. This PR fixes the
issue by copying the same logic from `skpro` `BaseProbaRegressor`.

The PR and estimator in https://github.com/sktime/sktime/pull/6424 can
be seen as a test. Fixes #7145.