[ENH] Added get_test_params in ThetaLinesTransformer  (#7199)
This PR adds the get_test_params method to the ThetaLinesTransformer
class, providing standardized parameter settings for testing.

This change implements the get_test_params method, which returns a list
of dictionaries containing test parameters for the
ThetaLinesTransformer. This will facilitate easier testing and
validation of the transformer’s functionality.