Update README.md
Sir, the logo or icon of the tool should be at the top. As you said, I have reduced the size of logo and replaced welcome to sktime with About. In this way, the links are also in good position and the header is also looking good. How do you think?