[ENH] New splitter switching from expanding window to sliding window at cutoff point (#7193)
#### Reference Issues/PRs

Fixes #7191 - Combining expanding and sliding window splitter in one


#### What does this implement/fix? Explain your changes.

This PR implements a combination of the sliding window splitter and
expanding window splitter. It will by default behave as the expanding
window splitter up to a point. After the maximum expanded window size is
reached, the splitter continues as a sliding window splitter.