[BUG] Fix `ExpandingCutoffSplitter` when `step_length` > 1 (#7290)
#### Reference Issues/PRs
Fixes https://github.com/sktime/sktime/issues/7289

#### What does this implement/fix? Explain your changes.
Fixed an "index out of bounds" error in the ExpandingCutoffSplitter
class that occurred when step_length was greater than 1 and used with
the evaluate function. The issue was due to incorrect handling of the
step size in the internal indexing logic, leading to attempts to access
invalid indices. Updated the logic to correctly manage index boundaries
under these conditions