[ENH] homogenization of `sktime` and `skchange` detection API - part 1 - base class and `y` argument (#7342)
This carries out the most substantial parts of issue
https://github.com/sktime/sktime/issues/7323 to homogenize the `sktime`
and `skchange` detection interface, with deprecation & change schedule
to ensure downwards compatibility and user warnings.

* `BaseSeriesAnnotator` is changed to `BaseDetector` in the
`sktime.detection` module, and in its downwards use. The
`BaseSeriesAnnotator`, in its current export locations
`sktime.annotation.base` and `sktime.detection.base` remains, but is
changed to a child of `BaseDetector` with an added warning.
* signature of `fit` and fit-like methods is changed to take `y` instead
of `Y`. As `y` takes the place of `Y`, making positional argument calls
downwards compatible. To ensure keyword calls are also compatible, the
argument `Y` is patched into `y` internally if passed, such a call
raises a warning. To ensure extension contracts remain compatible,
internally, we check which inner argument of `_fit` we need to pass `y`
to.