[ENH] `EnsembleForecaster` option to specify multiple copies of same forecaster easily (#7424)
Resolves #7407. 

A common use case with ML estimators is to create an ensemble estimator
with N replicates of the same estimator. This PR facilitates this use
case.

Previously, an ensemble forecaster was created via EnsembleForecaster(
forecasters, ... ), with forecasters being a list of 2-tuples [(str1,
estimator1), ..., (strN, estimatorN)]. This is a bit tedious if many of
the estimators are identical.

This PR allows each tuple to be either of the form (str, estimator) (as
before) or (str, estimator, count). This latter form says to create
count replicates of the estimator.