[ENH] homogenization of sktime and skchange detection API - base class `predict_points` return type (#7459)
#### Reference Issues/PRs

This change is part of the ongoing effort to homogenize the base class
detection module and ensure compliance across the BaseDetector
hierarchy, as listed in #7323.

#### What does this implement/fix? Explain your changes.

- Updates `test_predict_points` in the `test_all_annotators` suite to
enforce and validate the return type of `pd.Series` for the
`predict_points` method of `BaseDetector` as it's specified in its
contract.
- For `segmentation` tasks, this update also converts the return type of
`predict_segments(X)` in this method to `pd.DataFrame`, as that is the
expected parameter type to the `segments_to_change_points` call
  - `GreedyGaussianSegmentation` was previously passing `np.ndarray`
- This test is currently on the skip list for several failing estimators
with issues already open per estimator to fix this failing test among
others. The following estimators will need to update their `_predict`
method to adhere to the contract specified for this method. In
particular, segmentation tasks should have`_predict` return a
`pd.Series` with an index of intervals, which is also specified in the
`_predict_segments` method that calls `_predict` (to be addressed in the
attached issues)
  - GreedyGaussianSegmentation - #7412 
  - ClusterSegmenter - #7462