[ENH] Reduce memory footprint of `statsmodels` adapter - save necessary `_y` info to remove need for `_y` (#7417)
#### Reference Issues/PRs
Fixes #7405
Discussion #6914
The bigger issue: #7416

#### What does this implement/fix? Explain your changes.
It's not necessary to save all of `_y` and `_X`, but if memory conscious
and using `set_config(remember_data=False)` prediction methods will not
work. So saving only the necessary info to allow prediction.


#### What should a reviewer concentrate their feedback on?
- Saving the `_y` and `_X` info the way I am seems kind of arbitrary.
Should everything go in `_y_metadata` (or the respective attribute for
`_X` (if there is one)?