[DOC] Docstring for PanelGluontsList added (#7539)
Add short and long descriptions, conditions and capabilities in the
docstring for PanelGluontsList.

#### Reference Issues/PRs
Fixed docstrings for PanelGluontsList as per specifications in #7386.


#### What does this implement/fix? Explain your changes.
Added detailed description in documentation for PanelGluontsList. Also the older data type description "polars.DataFrame based specification of panel of time series." is clearly incorrect for PanelGluontsList (which is derived from the gluonTS library). It was likely carried/copied over from the preceding datatype PanelPolarsEager, which does have that description. Fixed that to have the correct description from
MTYPE_REGISTER.