[ENH] Forecaster imputer for RecursiveReductionForecaster (#7535)
#### Reference Issues/PRs
Fixes #7477 

#### What does this implement/fix? Explain your changes.
This proposal aims to introduce the capability to incorporate a
`forecaster` into the imputation process within the
`RecursiveReductionForecaster `(RRF). Specifically, it suggests allowing
RRF to accept a forecaster as an initialization argument and to
initialize the Imputer with` method="forecaster"` during the `fit`,
`predict_in_sample`, `predict_out_of_sample `when a forecaster is
provided. This feature would be a valuable enhancement, delivering
greater flexibility and precision in managing missing values.