[ENH] Validation split fix issue 7644 (#7647)
Description:

This update (Fixes #7644) modifies the code to ensure that the y_test
dataset is only created when validation_split is not None. If
validation_split is None, no y_test dataset is created, and it is not
passed to the Trainer. This change optimizes resource usage and
eliminates unnecessary processing when validation data is not required.

Key Changes:

Added a conditional check to create y_test only when validation_split is
provided.
Ensured that y_test is excluded from being passed to the Trainer when
validation_split is None.