[BUG] fix use of Imputer in RecursiveReductionForecaster (#7706)
#### Reference Issues/PRs
This is a minor bug fix discovered during usage, the bug was probably
introduced in the following PR:
https://github.com/sktime/sktime/pull/7646

#### What does this implement/fix? Explain your changes.
This fixes the use of a string instead of an imputer for lags by adding
a missing `_` to `self.impute_method`. The issue caused the imputer
functionality to fail when processing lags.