[ENH] windowed F1-score for detection (#7628)
#### Reference Issues/PRs
Fixes second half of this issue:
https://github.com/sktime/sktime/issues/7511 by implementing an F1-Score
class that inherits from BaseDetectionMetric (PR for the implementation
of this class can be found at
https://github.com/sktime/sktime/pull/7515/files).

#### What does this implement/fix? Explain your changes.
This class implements the computation of the F1-Score. It analyzes the
similarity between ground truth and predicted segments, returning a
float value between 0.0 and 1.0 (0.0 being no similarity, 1.0 being
perfect overlap).