[BUG] fix `all_estimators` lookup in case a tag is used that is not scitype specific (#7679)
Fixes https://github.com/sktime/sktime/issues/7664

The exact failure condition: `all_estimators` would ignore the type
specified via `estimator_types` if `filter_tags` was also provided, in a
case where all tags filtered for were available in multiple scitypes.

The reason was the internal translation code that translated the
`estimator_types` argument to a `filter_tags` argument or `object_types`
in `scikit-base` `all_objects`.

This has now been fixed, and a test has been added to cover
https://github.com/sktime/sktime/issues/7664.

An existing test for estimator type plus tag filter was present, but it
did not trigger the condition as the tag used did not occur outside
forecasters. This test has been extended to also use the tag triggering
#7664, `handles-missing-data`.

The test framework was also not affected, as the retrieval calls in test
collection were not using the `filter_tags` argument.

Also makes formatting improvements to the docstring.