[ENH] Implemented statmodels' `ar_select_order` (#7693)
#### Reference Issues/PRs
Solves part 2 of #4578


#### What does this implement/fix? Explain your changes.
A parameter estimator was added to estimate the best set of lags an
AutoREG model should use to achieve the best results.

#### What should a reviewer concentrate their feedback on?
Addition of private functions to _fit, I did not notice any other
implementation where there were private functions created inside _fit
and I want to ensure that it is right to do so. I also tried using soft
dependencies the way that are supposed to be used but I'd still request
a check to ensure I used them correctly.

#### Did you add any tests for the change?
Added a test to check if my implementation matches that of statsmodels
(without exog).