[ENH] `sklearn 1.6.X` compatibility patch (#7751)
Fixes some incompatibilities with `sklearn 1.6.X`:

* `test_FeatureUnion_pipeline` was testing compatibility of classifier
pipelines with `sklearn` components. This is no longer supported as
`sktime` has transitioned to own components, and `sklearn` made
compositors more restrictive with the "metadata routing" mechanism. The
test is therefore changed to `xfail` (not removed to have a reference
for potential pipeline redesigns).