[BUG] conversion `Series`-`pd.DataFrame` to `Series`-`pd.Series` now retains original column name as attr name in `pd.Series` (#7762)
Fixes https://github.com/sktime/sktime/issues/7763, the following:

Suppose XYZ is a forecaster with y scitype of pd.Series.
If you call XYZ.fit(y) with y a pd.DataFrame with a single column,
sktime is "robust" and will convert the single column DataFrame to a
pd.Series. This is done in routine 'convert_MvS_to_UvS_as_Series' in
datatypes/_series/_convert.py.

The problem is that in doing this, the column name from the DataFrame
has been lost. It should be retained as the attr name in the new series.
This PR modifies the behavior of convert_MvS_to_UvS_as_Series so that
the attr name of the new pd.Series object is the column name from the
original pd.DataFrame.

I ran pytest datatypes pre- and post- changes and it had the same number
of passes and skipped, and no failures. (Although there were some slight
changes on ~ 3 lines which I did not understand.)