[BUG] fix `load_fpp3` with fallback URL to avoid failure from changing CRAN URLs (#7799)
Yesterday, Felipe reported test failures coming from test_load_fpp3. Per
the design requirements for the fpp3 datasets, the data is not stored
within the sktime GitHub. Data is retrieved from a remote repo (in this
case, CRAN). The remote repo was updated and this required a
corresponding update on the sktime side. Fixed.