[ENH] `STLForecaster` - added method `plot_components` (#6423)
STL decomposition is an important and widely used technique in time
series analysis. It is supported in sktime via the STLForecaster class.
While it can be used as a step in a chain, it is also helpful - and
common - to plot the decomposed series - observed, trend, seasonal,
residuals. The method plot_components() has been added to the
STLforecaster class to support this.

For an example of how to use this see section 2 in the attached pdf.

[PullRequestNotes.pdf](https://github.com/sktime/sktime/files/15316079/PullRequestNotes.pdf)