[BUG] fix sporadic out-of-bounds error in `ProximityStump` and WDTW (#7836)
This PR fixes an out-of-bounds error in `ProximityStump` and WDTW.

The error is due to a sporadic out-of-bounds where, in the WDTW
distance, weights of form w(abs(i-j)) are copmuted, with w a
pre-allocated weight vector. The pre-allocation size takes only the
possible range of i into account, but not the range of j.

This is fixed.