[ENH] improve condition and error message in `test_excluded_tests_by_test` (#7904)
This PR improves, and fixes, the test `test_excluded_tests_by_test`,
whose sole purpose it is to remind contributors actioning one of the
estimtaors in https://github.com/sktime/sktime/issues/3429 to remove the
test skip from the estimator.

* the condition is changed so the test triggers only in this scenario,
and not when a new estimator is added that does not have enough test
parameter sets. Currently the intended condition and the latter as well
trigger a failure, and these triggers correspond to the two "subset"
relationships of the set equality. Removing one removes the undesired
trigger condition.
* The error message is updated to be clearer and more actionable to
contributors.