[ENH] Adds Seasonal dummies features via a transformer SeasonalDummies (one hot encoding for seasonal categorical variable) (#7915)
Time series models often capture seasonal effects. This PR makes it easy
to add the seasonal features as exogenous variables available to the
model, via a new sktime transformer SeasonalDummies().

There is no test included but the docstring contains an example of how
to run, which could be converted into a test.