[ENH] trigger differential tests if core framework modules change (#7923)
This PR tries to cover some blind spots in the testing framework.

It triggers differential tests for estimators and object where
previously they would not have been triggered, namely if one of the
three key framework modules changes: `datatypes`, `tests`, and `utils`.

This condition is added to the `run_test_for_class` switch.