[BUG] hack-fix for round trip conversions of `pandas` time series (#7925)
Towards https://github.com/sktime/sktime/issues/7924 - by catching the
case of "0" `name` explicitly.

This case normally does not arise except through conversions, but this
is not a perfect fix, as it does not address the underlying problem.

Meant to be a temporary revert of @ericjb's `name` fix until we have
fixed the estimators that were incorrectly relying on the old behaviour.
See https://github.com/sktime/sktime/issues/7928 for discussion.