[ENH] remove `numpy 2` dependency handler from `BaseObject` (#7955)
Half a year ago, we added the method `_handle_numpy2_softdeps` to handle
dependencies that were not `numpy 2` compatible, but did not advertise
that bound in their package dep bounds. The method was intended to be
temporary, until packages move to declare their `numpy` bounds properly,
or become compatible with `numpy 2`.

Most soft dependencies have moved to `numpy 2` compatibility, but
`pmdarima` has not, and perhaps will not put this in their dependency
bounds - it is the only such dependency remaining.

This PR removes the temporary utility and moves the bound into the
`pmdarima` adapter.