[ENH] add clusterer use to `GroupByCategoryForecaster` and docstrings (#7782)
This PR adds the possibility to pass a clusterer to
`GroupByCategoryForecaster`.

It also looks like the `GroupByCategoryForecaster` is mostly a copy of
`TransformSelectForecaster`, so I do not quite understand what happened
here.

In any case, I've tried to sync the classes as much as possible, with
various features that have since been added to
`TransformSelectForecaster`.