[ENH] in tests, handle soft dependency in `BaseForecaster.predict_proba` on `skpro` (#7981)
In the "by-module" tests, the following condition could lead to
unexpected failure:

something changes in two top level modules, and one of them is a
`forecasting` estimator with a valid, non-standard `predict_proba`
method - or one of the framework module.

The problem is that this combined change will lead to
`TestAllEstimators` being executed in the non-forecasting cluster of
tests, attempting to call `predict_proba` of the forecaster.

This requires `skpro` for representation of distributions, which is not
installed in the non-forecasting test VM, producing an exception.

The change ensures that the test for `predict_proba` is not run if
`skpro` is not present - it is present in the forecasting tests which
will always run.