[BUG] Fix `HierarchyEnsembleForecaster` test failure (#7992)
#### Reference Issues/PRs
Fixes test case failing for `HierarchyEnsembleForecaster`, found as part
of #7971.


#### What does this implement/fix? Explain your changes.

Test error that this pull request addresses:
`AssertionError('Estimator HierarchyEnsembleForecaster should not change
or mutate the parameter default from PolynomialTrendForecaster() to
PolynomialTrendForecaster() during fit.')`
In most places, the `default` forecaster is cloned, but the clone step
was ommitted in one location. This pull request rectifies that.