[ENH] extend `_check_soft_dependencies` to allow dependency disjunctions (#7943)
This PR extends functionality of `_check_soft_dependencies` to allow
dependency disjunctions, i.e., to specify "package X or package Y must
be present", with inclusive (logical) or.

To specify a logical "or", replace a package str with a list of strings.
On the second nesting level, this will be interpreted as "or".

There is no deeper nesting, i.e., only first ("and") and second ("or")
level nesting are allowed. By the distributive theorem, all logical
formulae in PEP 440 specifiers can be defined this way.

The PR also fixes two cases where this was requested:

* the disjunction between `darts` and `u8darts` in
https://github.com/sktime/sktime/issues/7228
* the disjunction between `mlflow` and `mlflow-skinny` in
https://github.com/sktime/sktime/issues/7940

Closes #7228, closes #7940