[BUG] Use delegatedforecaster in Hierarchical prophet, and add upper bound to Prophetverse version (#7980)
#### Reference Issues/PRs
Although this was not reason of the issue #7964, this issue allowed
detecting this fix.

#### What does this implement/fix? Explain your changes.
This PRs adds _DelegatedForecaster as base class for
HierarchicalProphet's placeholder, and bounds the version of
Prophetverse to <0.6.0.

Prophet 0.6.0 will change the hyperparameters (will simplify them, and
make them less ambiguous). So I'm planning to first bound the version,
then, in a future PR, update de docstrings and set the lower bound to
>=0.6.0.