[BUG] fix `Series`-to-`Primitives` transformer output index for hierarchical data (#8037)
In a combination of cases the `Series`-to-`Primitives` transformer would
not correctly handle the output data and drop the last index element
unexpectedly. This would happen if:

* a `Series`-to-`Primitives` transformer
* supports `Panel` and `Hierarchical` data natively
* and is passed a hierarchical `pd.DataFrame` to `fit_transform`

This is one of the root causes of
https://github.com/sktime/sktime/issues/8012, fixes
https://github.com/sktime/sktime/issues/8038

The fix is to catch the condition exactly. The reason for the failure is
that we need to deal with two cases separately:

* vectorization causes a new level of "all-zeroes" to appear. This
happens exactly if a `Series` is vectorized, because this would result
in an abstract "zero-level index", i.e., just a vector without index.
`pandas` cannot represent that, so it returns a 1-row-`DataFrame` which
has a row index containing the integer 0.
* in this case, we need to remove the superfluous level after
vectorization
* in all other cases - which were erroneously not dealt with (as they
did not exist in the code base so far?) there is no superfuous index
level to remove.