[BUG] Fix failing tests for SklearnClustererPipeline, TimeSeriesDBSCAN, ClustererPipeline (#8028)
#### Reference Issues/PRs
Fix failing tests for SklearnClustererPipeline, TimeSeriesDBSCAN, and
ClustererPipeline tracked in #8026. Bugs discovered via #7971.


#### What does this implement/fix? Explain your changes.
Updated the default implementation of predict_proba for clusterers to
return an array of shape (n_instances, n_clusters). This is the intended
shape, but the implementation was instead returning an array of shape
(n_samples, n_clusters).

Also updated the tests to allow for instances to not be assigned to any
cluster (indicated by cluster -1). This means that the probability
assigned for each instance may be less than 1.