[ENH] handle `BaseForecaster.predict_proba` `skpro` dependency in `TestAllEstimators` (#8023)
`BaseForecaster.predict_proba` depends on `skpro`, which can cause
"`skpro` not present" exceptions in `TestAllEstimators` run on a
forecaster when called in an environment that does not contain `skpro`.

In the current CI this happens only in the rare case when forecasters
are collected in a non-forecaster CI batch, it still should be handled
as these are unexpected failures.