[ENH] reducer - minor code clarity refactor (#7556)
This PR carries out a minor refactor to improve code clarity in the
`make_reduce` code tree.

The refactor moves two sizeable if/else branches into their own
functions, which clarifies which variables are used in which branch.

The result is troubling, as for instance `transformers` and `fh`,
`windows_identical` are not used in both branches.