[BUG] Allow `Merger` to work on hierarchical inputs (#8071)
#### Reference Issues/PRs

Fixes failing tests for `Merger` tracked in #8026.


#### What does this implement/fix? Explain your changes.

Merger is the only transformer that converts Panel data to Series. In
fact it is the only estimator with tag `{"scitype:transform-input":
"Panel"}`. When this was implemented, the hierarchical case was not
considered. This PR adds that support to fix the failing tests.

This PR also fixes a bug when `stride == 0`. The issue here was that in
changing from numpy3d to ndarray, the order of instances and time points
was not accounted for in this case.