[MNT] update `PyKANForecaster` dependency set (#8109)
The `PyKANForecaster` seems to require `matplotlib` now as one of its
dependencies, see failure below:

```
 FAILED sktime/tests/test_all_estimators.py::TestAllEstimators::test_fit_idempotent[PyKANForecaster-2-ForecasterFitPredictMultivariateNoX-predict] - ModuleNotFoundError: No module named 'matplotlib'
```

This is likely bad dependency management on the part of the package
maintainers - `matplotlib` is a plotting and not a modelling package -
but it cannot be helped as the import is somewhere on module level, so
it is a de-facto dependency now.

We need to keep track and add the dependency requirement, or the
estimator will break for users.