[ENH] Add simdkalman-based Series and Panel transformers (#8030)
This fixes https://github.com/sktime/sktime/issues/7688 and adds a soft
dependency to [simdkalman](https://simdkalman.readthedocs.io) (_EDIT:
initially this PR only had the Series version but now a minimal Panel
version is implemented as well_)

Simdkalman is most beneficial as a Panel transformer. However, it should
also work as a Series transformer, and this functionality is also
implemented here. This has the benefit that (some of) the existing
Kalman filter unit tests can be used to verify the functionality of the
package.

So far, most of the existing unit tests relied on the EM algorithm,
which is currently not fully implemented in `simdkalman` (only
estimating the noise covariances and initial state is supported). If
this is fixed there, more of the existing unit tests could be enabled
for all three implmenetations `KalmanFilterTransformerFP`,
`KalmanFilterTransformerPK`, `KalmanFilterTransformerSIMD`. For this
reason, some new unit tests where the results are compared to PyKalman
were also added.