[ENH] `ray` parallelization backend (#8031)
#### Reference Issues/PRs

Resolves: #8018


#### What does this implement/fix? Explain your changes.

Implements a direct parallel backend using ray. If ray is already
initialized it will use that, else it sets up a local ray instance. Also
supports setting a custom logger (i think you would maybe need that to
get the url of the local dashboard).

#### Does your contribution introduce a new dependency? If yes, which
one?

Adds ray >= 2.40.0 to the pyproject.toml in the all_extras and
pandas2_extras section