[BUG] fix `scitype` return in case `coerce_to_list=False` and `force_single_scitype=False` if two or more types are detected (#8133)
If `scitype` was called with `coerce_to_list=False` and
`force_single_scitype=False`, if two or more types were detected, a
single scitype would be returned - which is unexpected and inconsistent
with the docstring.

The test `test_scitype` also expected the single scitype output.

Given the strong expectation to the contrary, I think this should be
considered a bug.

However, since this is potentially breaking, we should release only with
a minor version, e.g., 0.37.0.

Depends on https://github.com/sktime/sktime/pull/8131 which is a pure
refactor.